backend_pipeline/requirements.txt
requests
python-dotenv
openai
moviepy
